Split ViewTest
